Refactor add domain, ui and io packages
