--- src/main/java/idle/game/generator/visualization/tool/GeneratorReader.java
@@ -1,6 +1,0 @@
-package idle.game.generator.visualization.tool;
-
-public interface GeneratorReader {
-
-  Generator readGenerator(String fileName);
-}
