--- src/main/java/idle/game/generator/visualization/tool/GeneratorWriter.java
@@ -1,6 +1,0 @@
-package idle.game.generator.visualization.tool;
-
-public interface GeneratorWriter {
-
-  void writeGenerator(Generator generator);
-}
