+++ src/main/java/idle/game/generator/visualization/tool/domain/io/GeneratorWriter.java
@@ -1,0 +1,8 @@
+package idle.game.generator.visualization.tool.domain.io;
+
+import idle.game.generator.visualization.tool.domain.Generator;
+
+public interface GeneratorWriter {
+
+  void writeGenerator(Generator generator);
+}
